requirements.txt
coverage==6.4.2
Django==4.0.6
djangorestframework==3.13.1
django_cors_headers==3.13.0
djangorestframework-simplejwt==5.2.0
